requirements-1.txt
asn1crypto==0.22.0
beautifulsoup4==4.6.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
cryptography==1.9
decorator==4.0.11
dryscrape==1.0
enum34==1.1.6
idna==2.5
ipaddress==1.0.18
lxml==3.8.0
mysql-connector==2.1.4
pycparser==2.18
pyOpenSSL==17.1.0
requests==2.18.1
selenium==3.4.3
six==1.10.0
urllib3==1.21.1
validators==0.12.0
webkit-server==1.0
xvfbwrapper==0.2.9
